requirements.txt
absl-py==0.11.0
backcall==0.2.0
block-extractor-cuda==0.0.0
cachetools==4.2.0
certifi==2020.12.5
cffi
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
dominate==2.6.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.3.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
jsonpatch==1.28
jsonpointer==2.0
kiwisolver==1.3.1
local-attn-reshape-cuda==0.0.0
lpips==0.1.3
Markdown==3.3.3
matplotlib==3.3.3
natsort==7.1.0
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
pandas==1.1.5
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow
prompt-toolkit==3.0.9
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
Pygments==2.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyWavelets==1.1.1
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
resample2d-cuda==0.0.0
rsa==4.7
scikit-image==0.17.2
scipy==1.5.4
six
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tifffile==2020.9.3
torch==1.0.0
torchfile==0.1.0
torchvision==0.2.1
tornado==6.1
tqdm==4.55.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.5
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.4.0
